front: fix-bug: set default standard allowance type to percentage instead of time (#3331)
Fixes #3151 

When the STDCM tab is opened from a fresh environment and the standard
allowance parameters aren't modified, the front sends a request with a
standard allowance of "0 seconds". This is invalid because it's a flat
time. It should be either unspecified or set to a percentage value.

The flat time option can't be manually selected anymore, but it's still
the default value.

I have modified the default valut to 0%.