editor: fixes switches port picking
This commit should fix issue #3394.

Details:
- Since hovered entities are not accurate enough, the hovered
track section is fully loaded before trying to determinate which
extremity is the closest to the mouse.
- The hovered endpoint is now saved in the SwitchEditionState, and the
exact same value is used when the user clicks, so that we are sure the
selected endpoint is the same as the one displayed over the map.