core: fetch and cache the infrastructure version
We previously mistakenly used the requested infrastructure version as
the reference for the current infrastructure version.

That is not correct, and cannot come close when the client does not request
a specific infrastructure version.

As the infrastructure is retrieved over multiple transactions, there
can be no guarantee whatsoever that the infrastructure is valid.

To mitigate this issue, the infrastructure version is retrieved first,
so that an up-to-date infrastructure can be downloaded later on.