core: kt-infra: stdcm: migrate availability legacy adapter (#4678)
Fixes #4265 

* Renamed the legacy classes to write new ones. `OccupancyBlock` has
been renamed to `OccupancySegment` to avoid ambiguity with infra blocks.
The legacy classes are still in use.
* `UnavailableSpaceBuilder` has been mostly rewritten to output block
occupancy. But the actual behavior remains unchanged for easier testing:
the aspect cascade is based on routes. This is technically wrong, but
behavior changes are avoided in this migration.
* `BlockAvailabilityAdapter` has been adapted to kt infra interfaces and
block elements.

**This version behaves in the same way as the previous one**, for easier
testing. This means that some computations are semantically wrong, such
as the spacing requirements. This class will be the replaced very soon
after the infra migration.