gateway: add an authenticating reverse proxy
Before this change, we relied on an authenticating reverse proxy called gateway. It's been in use since the inception of the project, as part of a package of services we got from SNCF Réseau.

The front-end was responsible for authenticating with keycloak, and providing a Bearer token for the reverse proxy to check.

This change introduces an authenticating reverse proxy, which replaces the previous gateway.
The front-end now entirely delegates authentication to the backend.

On startup, gateway/auth/login is called, which either returns an username
if the client is authenticated, or a redirection URL if the user needs to log in.

There are three authentication backends supported by the new gateway:

- A default mock backend, which return all users to be authentificated with the same
  username
- An OpenID Connect backend, which is used to authenticate interactive users.
  The application does not support session refresh, and does not forward logout events to the
  identity provider.
- A static Bearer token backend, which is not meant to be used by interactive users

This change also introduces unstable builds, which are pushed to ghcr.io.

Co-Authored-By: Valentin Chanas <anisometropie@gmail.com>
Co-Authored-By: =?UTF-8?q?=C3=89lys=C3=A6?= <101974839+ElysaSrc@users.noreply.github.com>