editoast: set the minimum tile level to 5
Currently, editoast as a big performance issue: generating the bigger
tiles takes a lot longer than we can afford, which blocks database threads
and can cause lockups.

As users open the app and zoom out, huge tiles need to be generated by the back-end.
A lot of these tiles aren't even useful, as the front-end will immediately request smaller
tiles, which don't cause as many issues.

This value is hard-coded as editoast does not have the performance budget
to generate bigger tiles.