editoast: pathfinding: add an upper bound compared to best solution
editoast computes k-shortest path. When we are along a simple line,
the next shortest path can be significant longer than the initial solution.
This commit limits any subsequent path to be at most twice as long as the best solution.

From an implementation point of view, we need to track the total length at each step.
This is a bit cumbersome, but we can’t access with the pathfinding crate to cost data.