editoast(refactor): renaming 'OperationResult' into 'CacheOperation'
`OperationResult`'s purpose is to apply operation on the `InfraCache` object.
The naming `OperationResult` come from the fact that `Operation` are applied on DB,
and the result of that operation is kept around as `OperationResult` to be reapplied
on the `InfraCache`.

However, with the implementation of the magic wand, we start to create operations that will
not be applied on the DB, but only on the `InfraCache`. Therefore, the naming `OperationResult`
doesn't make sense anymore.

Since the purpose is to apply these operations on the cache, hence `CacheOperation` name.