core: fix detection zone creation around crossing node
For a given crossing linking tracks North to South and East to West:
A train detected coming from North can only be on North or South
track (until detected again).
But the detection zone's role is to protect from conflicts, and there may be a
conflicting train from East to West.
So the detection zone of a node must comprise all track section endpoints
tied to the node.