core: regroup all conflicting check in one function
Checking for a conflict (spacing or routing) was done in two different
places in the code, once in a lambda (called conflicting), which was
only useful for routing conflicts (checking for different configuration)
and another one in the `detectRequirementConflict` were time overlap
was check.

This PR proposes to put the conflict check in one place.