core: remove 'openapi.yml'
The 'openapi.yml' file is seriously outdated and not maintained.
We do not have any automated way to detect changes needed for it
at the moment.

At best, it's misleading to keep it around.

Also, it might soon be completely pointless, since `editoast` and `core`
will start to interact through RabbitMQ and not HTTP anymore.