front: fix text wrapping in speed space chart settings
Add `white-space: nowrap` to prevent the text from spilling on
multiple lines. Drop the `position: absolute` on the button to
toggle settings, it's better to just rely on flexbox instead
and avoid hardcoding a margin-left on the the settings form.

While at it, slightly reduce the font size for the checkbox
labels (purely cosmetic and subjective change).

Closes: https://github.com/OpenRailAssociation/osrd/issues/7589