core: ignore uncommon error when forward coasting doesn't intersect
We used to throw `ImpossibleSimulationError`, but that is
incorrect, coasting to a stop is generally the result of the
coasting envelope not intersecting with the base envelope.
We can "handle" it in the same way as the same scenario in backwards
coasting, by returning null to keep the binary search going.