front: don't use i18n pluralization for hiding zero count
react-i18next's pluralization feature requires providing at least
two variants: "_one" and "_other". An additional "_zero" variant
may be provided, but doesn't change the fact that "_one" is
required. This rule is broken by "manageVias".

One way to fix this would be to just add the variant by copy-pasting
"manageVias_other", however pluralization is designed to handle
one vs multiple, not zero vs one-or-multiple. Indeed, the choice to
hide/show the number of vias will never depend on the language. So
it doesn't feel like the right tool for the job here.

Instead, use a conditional in the react component.