front: fix default side value for signals in editor
The editor forms are automatically generated by rjsf from the infra
JSON Schema. The schema is automatically generated from Python
classes by pydantic. The schema contains some unnecessary "allOf"
fields (which could just be inlined), and rjsf chokes on them [1]
when handling default values. Unfortunately the "allOf" fields
cannot easily be simplified (because of the auto-generation),
and rjsf cannot easily be fixed (the logic to handle defaults is
quite involved).

As a workaround, manually specify the default for extensions.sncf.side
so that the form data always includes this field.

Closes: https://github.com/OpenRailAssociation/osrd/issues/7445

[1]: https://github.com/rjsf-team/react-jsonschema-form/issues/3599