editoast: use default value for missing SignalSncfExtension::side
The JSON Schema has this field marked as optional. Instead of
rejecting requests which don't include the field, fill the field
with the default value.

This still isn't enough to fix the original bug, because the form
doesn't reflect the correct default (shows "LEFT" instead of
"CENTER" initially, even if the field is undefined). But I believe
it's still worth fixing editoast.

References: https://github.com/OpenRailAssociation/osrd/issues/7445