front: fix crash on plaintext HTTP error response
When editoast is down, the gateway returns a 502 with a
text/plain error message. Example:

    HTTP/1.1 502 Bad Gateway
    content-type: text/plain; charset=utf-8

    Failed to connect to host: Failed resolving hostname: failed to lookup address information: Name or service not known

rtk-query is bubbling up this error as an object with "status"
and "data" fields, but with a string in the "data" field
instead of an object:

    {
        "status": 502,
        "data": "Failed to connect to host: [...]"
    }

The getErrorMessage() helper was choking on such values, trying
to cast "data" to an object unconditionally. Add a check to
ensure that the "data" field is an object before casting. If
"data" is a string, use that as-is for the error message.

(Unfortunately, TypeScript doesn't offer a way to check whether
an arbitrary value satisfies an interface, which would've been
handy here.)

While at it, rejigger the function logic a bit to simplify and
leverage early returns.

Closes: https://github.com/OpenRailAssociation/osrd/issues/7319