docker: fix SIGTERM not received by editoast
For some reason it seems like SIGTERM is not properly sent to
editoast when bringing down the cluster. I see the following in
my logs after a timeout:

    WARN[0010] StopSignal SIGTERM failed to stop container osrd-editoast in 10 seconds, resorting to SIGKILL

It seems like sh is the cause of the issue somehow. Use "exec" to
replace the sh process and fix the Unix signal issues.

With this patch, the shutdown is instantaneous and I now see the
following in the logs:

    [editoast]     |   2024-06-20T12:37:41.520662Z  INFO actix_server::server: SIGINT received; starting forced shutdown

(Maybe this is all specific to my Podman-based setup? I still think
it's worth fixing.)

While at it, use "&&" to abort startup if the diesel migration fails.