front: use noreferrer for external license links
When clicking a link to an external website, the Web browser might
send the URL of the OSRD webapp in the external HTTP request. The
URL might contain potentially sensitive information such as a
private/internal domain name or URL parameters. Make the Web browser
always omit this information by using rel="noreferrer".

No need for noopener here like #7828 did, because the link is opened
in the same window (we aren't using target="_blank").