readme: add missing copyright notice
It seems like a copyright notice [1] is missing from the project.
The copyright notice has a legal significance in most countries,
so it's safer to include it to make it clear that the work is
copyrighted.

The copyright is shared by all contributors to the project, reflect
that in the legal author.

[1]: https://en.wikipedia.org/wiki/Copyright_notice