front: fix chart synchronizer eco data
The useSimulationResults() hook is responsible for feeding train
simulation data to ChartSynchronizerV2. ChartSynchronizerV2 is
used to populate the top banner with the current train state (head,
tail, speed, and so on) and the graph pointers (vertical trace with
small dot).

The simulation data returned by the backend contains two fields:
"base" and "final_output". "base" ignores margins, "final_output"
takes them into account. In frontend parlance, "eco" is equivalent
to "final_output".

Before this patch, the base and eco arrays are built in a single
loop. This doesn't work because "base" and "final_output" may
contain arrays with different lengths. In other words, it's
incorrect to use an index referring to "base" and use it to access
"final_output": these entries don't refer to the same instant in
time.

This causes two issues: first, we're skipping the last entries from
"final_output" ("base" is shorter than "final_output" in practice).
Second, we're mixing up times, positions and speeds which are
unrelated. For instance, we use the time coming from "base" and mix
it with a position/speed coming from "final_output".

Fix these issues by splitting the loop into 2, one responsible for
"base" and the other responsible for "final_output".