front: fix unnecessary calls to simulation_summary by making postV2TrainScheduleProjectPath a query.
Codegen makes all the POSTs mutations by default. This one is only a POST to have a body, but is not creating/modifying anything in the DB.
No reason to be considered a mutation on the frontend.

As a query it doesn't have an invalidatesTags array anymore and prevents the unnecessary calls to simulation_summary.