front: add simulation result stops output table
This commit implements a new component to display the simulation results in a summary table.
Stops, margins, arrival time request before the simulation are displayed, and compared
with the actual stop, margins, times we got during the simulation.
The table will highlight the train stops where too much discrepancy is observed.