front: fix power restriction ranges edition
When creating a new pathStep while adding a power restriction range, the pathStep is located from its
 position on the path. Thanks to it and the cumulative sums of the track ranges' lengths on the path,
 we can find the track range on which it is located and its offset within the track range, and then its
offset within the track section (especially if the track range does not cover the whole track section).

The computation to find the offset used an incorrect cumulative sum, which led to an incorrect offset.