front: convert track offset to meters when stashing in the store
The backend would choke with "java.lang.RuntimeException: The
waypoint is not located on the track section XXX" when creating
a train with an itinerary defined by clicking on tracks on the
map, and then editing that train without changing anything (edit
then save immediately).

bc3c0028ca48 ("front: send track offset in mm in train schedule v2
requests") has modified our logic when sending path items to the
backend to convert meters to millimeters. However, when receiving
path items from the backend, we need to convert millimeters back
to meters as well. Adjust useSetupItineraryForTrainUpdate() to
perform this conversion.