core: fix node relative time diff calculation
Relative time diff calculation should return the lowest possible relative time diff in the interval made of the possible times at which the train can pass through the node. I.e. it should return 0 if it is possible to add delay for the train to pass at the planned arrival time.