front: upgrade NGE to forked v2.7.2
We maintain a fork of NGE with some custom patches:
https://github.com/osrd-project/netzgrafik-editor-frontend/tree/standalone

Our fork includes patches that aren't yet ready to be upstreamed,
or cannot be upstreamed at all. For instance, hiding the return
travel for trains (making them one-way) will require a much larger
refactoring to be supported upstream. Some features like saved
filters could in theory be supported but as-is will be lost on page
reload.

Unfortunately this means we're back to the @osrd-project namespace
for now.