front: fix editing a train add stops at timed points
When a user creates a train with a scheduled waypoint, the frontend sends the schedule of the path step
with a stopFor undefined, but the backend returns the schedule with a stopFor null.
The condition `isNaN(Number(step.stopFor))` is then true for a stopFor null, even if we have not modified
the path step, and the waypoint is transformed into a stop.

To prevent this, I added a new condition `isNil(step.stopFor)` which checks that the stopFor value is neither
undefined or null.