front: fix train edition in input table
when the list of suggested Operational points contained 2 points with the same uic but different ch, the function was not checking for ch and thus was choose the wrong op.

because of this the input table was displaying the data of some lines on wrong lines

now fixed, and some tests.