front: drop hardcoded scenario header height, second take
This is a re-do of #7993, which got reverted in #8015 because of a
bug when the timetable contains lots of trains.

In a flex context, the implicit default value for min-height is
auto, and that prevents elements from overflowing with a scrollbar
(because min-height wins against height and max-height). Override
the default with "min-height: 0".