front: pass train schedules as argument to loadTrainIds()
There is a race between the /train_schedule endpoint (used to
retrieve all train schedules) and the /train_schedule/simulation_summary
endpoint (used to retrieve a batch of simulation summaries).
Depending on the order in which responses are received, the train
might be dropped. If /train_schedule/simulation_summary comes first,
formatTrainScheduleSummaries() will not have the train schedule
to work with, thus will ignore the train.

Fix this by passing a list of train schedules to loadTrainIds().
The function can retrieve simulation summaries and directly has
up-to-date train schedules to work with.

Rename the function to loadTrains() since it no longer takes IDs.