front: replace some type assertions with non-null assertions
In general I prefer to avoid `as` type assertions as much as
possible, because it makes the compiler ignore any type mismatch.
The non-null assertion operator `!` is a bit better: if the types
don't match, the compiler will error out.

Use it in a bunch of places.

See https://github.com/OpenRailAssociation/osrd/pull/8698#discussion_r1745144151