editoast: provide search object list to the store explicitely
Search object configurations used to be collected in a static list for each
`Search` macro invocation. This list is then used in `derive(SearchConfigStore)`
to generate the store logic.
On top of being a Rust antipattern, it now triggers some warnings on
nightly 1.83.0 about mixing immutable and mutable references on static
data (which is indeed a bad idea). This static list also confuses
rust-analyzer. Also getting rid of it removes our only use of unsafe.

Signed-off-by: Leo Valais <leo.valais97@gmail.com>