editoast: locked rolling stocked returns a 409
When a rolling stock is locked and we try to update it,
we return a [400 BAD_REQUEST](https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.1).
I don't think the request is malformed in case of a locked
rolling stock.

Based on the description of what is a 409, the behavior
of our system in case a rolling stock is already locked
is more a [409 CONFLICT](https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.8).

Signed-off-by: Jean SIMARD <woshilapin@tuziwo.info>