front: prevent the via card to be demounted and remounted several times
The via card was remounted because its key changed at each re-render (the key was nextId()).
To avoid this, I modified the behaviour to create a via and give by default an id and a uic (= -1).
This way, we can use the via.id for the key. It solves the main issue.

To handle the "fake" uic, I created a new hook usePathfindingIsPossible which filter the invalid uic and launch
a pathfinding.
This hook is simpler than the existing usePathfinding hook, but I had to do some refacto in the map props (to give the path
geometry separately from the path properties). This will be useful for a later refacto.

Signed-off-by: Clara Ni <clara.ni@outlook.fr>