front: replace any with unknown in generated OpenAPI code
`any` is problematic because it spreads to the rest of the codebase
and is annoying to unwrap. Also it causes warnings at codegen time:

    /home/simon/src/osrd/front/src/common/api/generatedEditoastApi.ts
      1627:20  warning  Unexpected any. Specify a different type  @typescript-eslint/no-explicit-any
      2022:10  warning  Unexpected any. Specify a different type  @typescript-eslint/no-explicit-any
      2034:10  warning  Unexpected any. Specify a different type  @typescript-eslint/no-explicit-any
      2057:10  warning  Unexpected any. Specify a different type  @typescript-eslint/no-explicit-any

Fix this by turning on the `fixToUnknown` rule for `no-explicit-any`
to automatically convert any `any` type to `unknown`.

Ideally the codegen should be fixed to use `unknown` instead of
`any`, see:
https://github.com/oazapfts/oazapfts/issues/676

Signed-off-by: Simon Ser <contact@emersion.fr>