front: stop matching KP in matchPathStepAndOp()
In many spots we need to match a PathStep with a SuggestedOP.
Each spot is a special snowflake and needs its own little
variation of the matching logic. formatSuggestedViasToRowVias()
needs to match KP and PathStep.id, but that's undesirable for
other spots.

Extract the extra special matching near the spot it's needed.
In following commits matchPathStepAndOp() will be re-used in more
spots.

No behavior change, only refactoring.

Signed-off-by: Simon Ser <contact@emersion.fr>