editoast: update temporary speed limit import to use signals
Previously, the temporary speed limits creation endpoint expected to
retrieve directional track ranges in the request body. Since it might
not be possible to get that information directly, update the
endpoint so that it retrieves the list of execution and resume signals
associated with the temporary speed limits and then do a graph traversal
on the infrastructure track sections to determine from the signals on
which the track ranges the temporary speed limits apply.

Signed-off-by: Loup Federico <16464925+Sh099078@users.noreply.github.com>