core: fix "negative" routing requirements
Begin of routing requirement was sometimes placed later than the end
requirement of a zone.
This would happen when a stop on closed signal was after the considered
route (and this route was not the first one, as there is an early return).

Too early entrySignalOffset had to be fixed too.

Signed-off-by: Pierre-Etienne Bougué <bougue.pe@proton.me>