front: remove react-css-transition
This library didn't have a new release since 2022. It uses a react tool that will be deprecated in react v19 (findDOMNode).
It was used to display tooltips when hovering some buttons in the app.
Since the design system doesn't have a custom tooltip for buttons, it has been requested to just use the regular title html attribute.
Instead of finding a substitute lib, just remove the old one and use the title attribute.

Signed-off-by: SharglutDev <p.filimon75@gmail.com>