front: remove react-beautiful-dnd
The lib hasn't been updated since 2022 and now is outdated regarding React 19 new version.
The lib uses React "defaultProps" feature which is deprecated in React 19.
As the drag and drop feature isn't used by users and can be bypassed, remove the lib and the related feature to enable a future migration to React 19.

Signed-off-by: SharglutDev <p.filimon75@gmail.com>