front: fix some trigrams are not recognized
Since the search endpoint response is paginated, when the user searches a short string, the search endpoint will return
some of the operational points which match either the trigram or the name. As the search endpoint response is not ordered by trigram
first, we could miss some responses, and especially the operational points corresponding by trigram.

Now, if the user searches a short string, we will first get all the operational points matching the trigram and then the ones matching the name.
This way, we won't miss the reponse by trigram anymore.

Signed-off-by: Clara Ni <clara.ni@outlook.fr>