front: drop formatToIsoDate()
This is now unused.

This function was a bit weird, because it had two unrelated
purposes: format a human-readable string with a fixed frenchy
format, or format a machine-readable string for API purposes.
The function also accepted strings and numbers as input
(implicitly converting these to dates), making it so callers are
less type-safe.

Signed-off-by: Simon Ser <contact@emersion.fr>