front: fix isEqualDate()
toLocaleString() returns an arbitrary human-readable string. For
instance, in France it returns "18/02/2025 13:32:08" and in Korea
it returns "2025. 2. 18. 오후 1:32:08".

The string cannot be used for anything else than display. Trying
to extract the date from it will not work: substring(0, 10) will
give "2025. 2. 1" in Korea (dropping the "8"). As a result, any
day with the same first digit will be interpreted as equal.

Fix this by comparing the year, month, and day of the month instead.

Signed-off-by: Simon Ser <contact@emersion.fr>