front: drop TimeButtons, TrainDetails and ChartSynchronizer
These are unused at the moment, because the whole thing is hidden.
The code is bit-rotting and holding off refactoring.

Let's just nuke all of that. Maybe this feature will be re-introduced
at some point, but we'll need to rewrite big chunks of it anyways
and we can easily copy back useful bits from the Git history.

Signed-off-by: Simon Ser <contact@emersion.fr>