front: fix useEffet warning in useScenarioData
The call to get the train schedule results was recently moved into a useQuery.
By defaulting the respone with an empty array and using it in some useEffect dependencies, the useEffects were called multiple times (since the ref of the array changes at each rerender) while the endpoint was loading.

Signed-off-by: SharglutDev <p.filimon75@gmail.com>