front: cut down ESLint watch list
When ESLint starts up, it sets up a recursive file watcher on all
files/directories passed as arguments, regardless of
ignorePatterns. Passing "." causes performance issues because it
makes ESLint look at all files.

Instead, go back to the old approach: specify a list of files on
the CLI instead of ".".

Signed-off-by: Simon Ser <contact@emersion.fr>