front: fix weird behavior when adding train settings
This extra useEffect introduced a bug in end to end tests. The only way to reproduce it was to set the train name and start time in less that 500ms (the debounce used for these fields) which the e2e tests do.
At this speed, after filling the name, a useEffect is fired after 500ms after to update the store, but since the test is also filling the start time a few ms later, the useEffect was called because nameFromStore had been updated resulting in setting the start time back at its original value.
A workaround had been done for tests 011 and 012 to insert the selection of the rolling stock between the name and start time filling.

Signed-off-by: SharglutDev <p.filimon75@gmail.com>