front: avoid blinking with example DTO when launching NGE
NGE will use an example DTO (with Swiss train stations) if it's
initialized without a specific DTO. Later, we set the DTO to the
final state. There is a small time window where NGE displays the
example data before the real data. This results in blinking.

Set NGE's initial DTO as an empty state to avoid this blinking.

Requires this NGE patch:
https://github.com/SchweizerischeBundesbahnen/netzgrafik-editor-frontend/pull/419

Signed-off-by: Simon Ser <contact@emersion.fr>