front: fix console warning at first load of a non empty timetable
When loading the timetable of a scenario for the first time, while the timetable items are loading, we first displayed the checkbox corresponding to an empty timetable and then we switched to the other one.
As the initial checkbox had checked property non defined, React displayed the "uncontrolled" warning because he considered that the same checkbox was switching from an undefined state to a defined one.
Fix it by merging the two checkboxes.

Signed-off-by: SharglutDev <p.filimon75@gmail.com>