front: extract keys with TypeScript
i18next-parser uses the TypeScript compiler only for lexing. In
other words, it has no contextual information about the translation
calls. This makes extracting translation keys more fragile and
error-prone. For instance, when TFunctions are passed around, the
namespace is lost. As another example, "i18n.t" is not recognized
as a translation function. In general, this is a fool's errand:
operating on tokens without type information will always result in
a lot of unhandled corner cases.

Instead, use the TypeScript compiler API to extract type information
and iterate over all calls to the translation function.

Signed-off-by: Simon Ser <contact@emersion.fr>