ui-charts: adds new TrackOccupancyStandalone
This commit basically replaces the previous
trackOccupancyDiagram/rendering story, with a new component,
TrackOccupancyStandalone, that actually calls
useManchetteWithSpaceTimeChart instead of mimicking it. The main benefit
is to get proper interactions, such as panning on the time axis.

Also, this commit is very impactful, because many things were done quite
customly for track occupancy drawing, when the SpaceTimeChart APIs
provide many tools to help rendering additional layers on top of them.

One final note, before getting into the detail: This commit breaks
selecting a train with the mouse on the TrackOccupancyDiagram, because
this was done in the rendering loop, rather than using picking (as it
should).

Details:
- Replaces all time and space to pixels conversions from track occupancy
  layers, with code that uses proper tools from SpaceTimeChartContext
- Removes mouse collision detection (it was done in the rendering
  process, which meant that rendering was depending on the
  mousePositions, causing many excess renderings)
- Updates OccupancyZone: id becomes trainId, and arrivalTime and
  departureTime become numbers instead of dates
- Adds new TrackOccupancyStandalone component
- Updates rendering story to use new component

Signed-off-by: Alexis Jacomy <alexis.jacomy@gmail.com>