front: fix open API import date timezone
We used UTC instead of the local timezone. As a result, the date
would be wrong when around midnight in France (UTC+2). The date
would be wrong more often in timezones with a bigger UTC offset
(e.g. in Korea, UTC+9).

Fix this by replacing formatIsoDate() with a helper similar to our
existing formatLocalDateTime() function.

Signed-off-by: Simon Ser <contact@emersion.fr>