editoast: use const generics to declare maximum and default page size
Introduces a subtle API breaking change: now the default page_size is
its maximum value. Before, we could have a default different (but lower)
than the maximum.

Rationale: if the maximum is correct, the amount of data queried and
sent back is reasonable. Therefore using that as a default is fine
and reduces the amount of requests we recieve. On the other hand, if
that `MAX_PAGE_SIZE` value is too high as a default value, it's likely
plain and simply too high.

Signed-off-by: Leo Valais <leo.valais97@gmail.com>