editoast: simplify PaginationQueryParams interface and error management
Paginations errors will be treated as serialization errors just like our
schemas.  That means they won't have a specific key to match for
translations in the frontend.  But since this is never a user error but
rather an API caller error/bug, it probably doesn't need translation
since it won't mean anything to the end user.

Signed-off-by: Leo Valais <leo.valais97@gmail.com>