editoast: remove rs inertia coeff uom ratio typing
Uom ratio supports conversion between different ratio units, such as percentages, per thousand, per one...
But inertia coeff is a pure dimensionless quantity for which unit conversion never makes sense,
and it does not represent a count of something per something else as the uim ratio units suggest.
Thus I believe it makes sense to keep it as a pure number.

This is compatible with uom computations,
and .into checks that the quantity is indeed dimensionless to authorize into f64,
so we should not lose type safety.

Signed-off-by: Alice K. <alice.khoudli@gmail.com>