editoast: fix `/authz/me/grants` for admins
Confusion in the Regulator between privileges and grants.  The `Admin`
role simulates fake privileges, but must not make up grants.

The endpoint response, when queried by an admin, now doesn't return any
grant if the admin doesn't have one.  The admin is still able to
use the resource (by design).

Only the checks have been removed, but these functions will be merged
later so the Regulator won't stay in this state for long.  Let's fix the
bug asap nonetheless.

Signed-off-by: Leo Valais <leo.valais97@gmail.com>