editoast: fga: customize `impl Type` parsing
Sets up the ability for types implementing `fga::model::Type` to have a
parsing step when recieving identifiers from OpenFGA (like
`mytype:myid`).  This will be useful to have more convenient model
types.

The primary goal is to be able to use `enum Role { Admin, Stdcm, .. }`
with OpenFGA queries directly.  Instead, we currently have a `struct
Role(String)` used as a way to communicate with the OpenFGA client and
we need an extra parsing step.

We'll also be able to have `User`, `Group` and `Infra` store an `i64`
representing the DB ID directly, instead of converting into and from a
String.

Last but not least, we'll be able to have much stronger `Authorizer` and
`Regulator` signatures.  That'll lower the chances of confusing numeric
IDs in function calls.  The goal is to have:

    pub async fn user_groups(&self, user: &User) -> Result<HashSet<Group>, Error<S::Error>>;

instead of (like today):

    pub async fn user_groups(&self, user_id: i64) -> Result<HashSet<i64>, Error<S::Error>>;

Signed-off-by: Leo Valais <leo.valais97@gmail.com>