front: only fetch train list once when opening scenario
We were using useQuery(), which refetches data automatically when
the rtk-query cache is invalidated. This causes issues when turning
a paced train into a train schedule (or the other way around):
deleting the old paced train triggers a refetch of the paced train
list and overwrites timetableItems with a list missing the new
train schedule.

Since we're maintaining our own timetable item list, ensure we only
fetch train schedules and paced trains once when opening a scenario,
and then rely only on upsertTimetableItems/removeTimetableItems to
keep our list up-to-date.

This also suppresses unnecessary requests.

Signed-off-by: Simon Ser <contact@emersion.fr>