editoast: create editoast_core crate and remove core module
Co-authored-by: Jean SIMARD <woshilapin@tuziwo.info>
Signed-off-by: hamz2a <atrari.hamza@gmail.com>

editoast: use 'testing' feature for 'editoast_core' in tests

Running tests from inside `editoast_core`, the resolver of `cargo`
does not have the `fixture` feature, and therefore the tests can't
compile. In the whole workspace, it works because the resolver has
the feature for some other packages in the workspace.

Signed-off-by: Jean SIMARD <woshilapin@tuziwo.info>

editoast: use 'mocking_client' for 'editoast_core' testing

Some of the tests of `editoast_core` needs the `MockingClient`.
However, the feature is not activated by default: when running the
tests inside `editoast_core`, compilation fails.
In the entire workspace, it works correctly because the `cargo` resolver
has other packages that need this feature.

Signed-off-by: Jean SIMARD <woshilapin@tuziwo.info>

editoast: add a doc test for how 'AsCoreRequest' works

Signed-off-by: Jean SIMARD <woshilapin@tuziwo.info>